Connected delivery with backend and modified backend logic

+++ server/common-data-service/src/main/java/com/erp/common/model/delivery/DeliveryModel.java
@@ -13,11 +13,11 @@ import lombok.NoArgsConstructor;
 public class DeliveryModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
     private String code;
-    private Integer value;
+    private long value;
     private String name;
-    private Integer population;
-    private Integer density;
-    private Integer userId;
+    private long population;
+    private long density;
+    private long userId;
 }
