+++ server/common-data-service/src/main/java/com/erp/common/repository/FaqRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface FaqRepository extends JpaRepository<FaqModel, Integer> {
+public interface FaqRepository extends JpaRepository<FaqModel, Long> {
     Optional<List<FaqModel>> findAllByUserId(long userId);
 
     @Transactional
